api/requirements.txt
fastapi>=0.115.11
fastparquet>=2024.11.0
google-genai>=1.3.0
numpy>=2.2.3
pandas>=2.2.3
pydantic>=2.10.6
python-dotenv>=1.0.1
tabulate>=0.9.0
uvicorn[standard]>=0.34.0